Separate api client into multiple files
